frontend: Move FlagsSummary to common code.
