+++ packages/frontend-web/src/app/components/LazyLoadComment/LazyLoadComment.tsx
@@ -29,7 +29,6 @@ import {
   ITopScore,
 } from '../../../types';
 import { REQUIRE_REASON_TO_REJECT } from '../../config';
-import { FlagsSummary } from '../../scenes/Comments/components/FlagsSummary';
 import {
   articleBase,
   commentRepliesDetailsLink,
@@ -50,6 +49,7 @@ import { CommentText } from '../CommentText';
 import {
   ConfirmationCircle,
 } from '../ConfirmationCircle';
+import { FlagsSummary } from '../FlagsSummary';
 import {
   MoreVerticalIcon,
   ReplyIcon,
